fix(RefreshToken): 토큰으로 member가 확인되지 않는 오류 수정
sping security 설정 추가

+++ src/main/java/com/example/ProJectLP/domain/vinyl/Vinyl.java
@@ -35,7 +35,7 @@ public class Vinyl extends TimeStamped {
     private String imageUrl;
 
     @Column
-    private Integer releaseYear;
+    private Long releaseYear;
 
     //추가되어야할 항목
     //좋아요
