+++ src/main/java/com/example/ProJectLP/domain/vinyl/VinylRepository.java
@@ -2,6 +2,6 @@ package com.example.ProJectLP.domain.vinyl;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VinylRepository extends JpaRepository<Vinyl, Integer> {
+public interface VinylRepository extends JpaRepository<Vinyl, Long> {
 
 }
